requirements.txt
python-dotenv
uvloop>=0.12.2
discord.py>=1.0.0

